+++ src/main/java/fish/common/community/comment/index/controller/CommentController.java
@@ -1,8 +1,8 @@
-package fish.common.community.comment.controller;
+package fish.common.community.comment.index.controller;
 
-import fish.common.community.comment.dto.request.CommentRequest;
-import fish.common.community.comment.dto.response.CommentResponse;
-import fish.common.community.comment.service.CommentService;
+import fish.common.community.comment.index.dto.request.CommentRequest;
+import fish.common.community.comment.index.dto.response.CommentResponse;
+import fish.common.community.comment.index.service.CommentService;
 import fish.common.user.entity.User;
 import fish.global.util.ResponseUtil;
 import lombok.RequiredArgsConstructor;
@@ -40,11 +40,5 @@ public class CommentController {
     public ResponseEntity<ResponseUtil> deleteComment(@PathVariable Long commentId) {
         commentService.deleteComment(commentId);
         return ResponseEntity.ok(ResponseUtil.success());
-    }
-
-    @PostMapping(value = "/comments-like/{commentId}")
-    public ResponseUtil<?> save(@PathVariable Long commentId, @AuthenticationPrincipal User user) {
-        commentService.saveCommentLike(commentId, user.getId());
-        return ResponseUtil.success();
     }
 }
